feat: refatoracao-home
