+++ src/pages/Home/provider.tsx
@@ -1,0 +1,15 @@
+const cards = [
+  {
+    title: "Trabalho",
+    description: "Notas para controle e distribuição de tarefas do trabalho",
+    link: "desktop",
+  },
+  {
+    title: "Faculdade",
+    description:
+      "Notas para controle e distribuição de tarefas de um trabalho da faculdade",
+    link: "desktop",
+  },
+];
+
+export default cards;
