+++ src/pages/Home/styles.tsx
@@ -1,0 +1,13 @@
+import { makeStyles } from "@mui/styles";
+
+const useStyles = makeStyles({
+  container: {
+    marginTop: "30px",
+  },
+  mainTitle: {
+    fontWeight: "200",
+    fontSize: "2em",
+  },
+});
+
+export default useStyles;
